Puesta producción segura/SecretosVariablesEntorno/src/requirements.txt
fastapi[standard]
uvicorn
pydantic
pyjwt
motor
passlib